# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.2.1"
  description: "API made for the SHCITY project"
  title: SHCity API
  contact: 
    name: "Alexandre Dias"
    url: "https://www.linkedin.com/in/alexandredias95/"
    email: "alex.m.a.dias95@gmail.com" 
# during dev, should point to your local machine
host: 193.136.127.202:80
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /occupation/areas_and_buildings_top_visits_add: 
    x-swagger-router-controller: area_and_buildings
    post:
      summary: Add the top on visits (Type 0 = Building; Type 1 = Area).
      tags:
              - occupation
      consumes:
        - application/json
      operationId: areas_and_buildings_top_visits_add
      parameters:
        - in: body
          name: Top_visits
          description: The Top_visits to add.
          schema:
            $ref: "#/definitions/Top_visits"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"
  /occupation/areas_and_buildings_level_occupation_add: 
    x-swagger-router-controller: area_and_buildings
    post:
      summary: Add the occupation (Type 0 = Building; Type 1 = Area).
      tags:
              - occupation
      consumes:
        - application/json
      operationId: areas_and_buildings_level_occupation_add
      parameters:
        - in: body
          name: Occupation
          description: The Occupation and waiting time to add.
          schema:
            $ref: "#/definitions/Occupation"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"         
  /occupation/areas_and_buildings_level_occupation_get: 
    x-swagger-router-controller: area_and_buildings
    get:
      summary: Get the occupation (Type 0 = Building; Type 1 = Area).
      tags:
              - occupation
      consumes:
        - application/json
      operationId: areas_and_buildings_level_occupation_get
      parameters:
        - in: query
          name: name
          description: The building or area to receive the value.
          required: true
          type: string
        - in: query
          name: type
          description: (Type 0 = Building; Type 1 = Area).
          required: true
          type: string   
      responses:
        200:
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"
  /occupation/areas_and_buildings_waiting_time_get:
      x-swagger-router-controller: area_and_buildings
      get:
        summary: Get the average waiting time (Type 0 = Building; Type 1 = Area).
        tags:
                - occupation
        consumes:
          - application/json
        operationId: areas_and_buildings_waiting_time_get
        parameters:
          - in: query
            name: name
            description: The building or area to receive the value.
            required: true
            type: string
          - in: query
            name: type
            description: (Type 0 = Building; Type 1 = Area).
            required: true
            type: string
        responses:
          200:
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/HelloWorldResponse"
          503:
            description: Error
            schema:
              $ref: "#/definitions/HelloWorldResponse"
  /occupation/areas_and_buildings_level_occupation_schedule_add: 
    x-swagger-router-controller: area_and_buildings
    post:
      summary: Add the occupation schedule (Type 0 = Building; Type 1 = Area).
      tags:
              - occupation
      consumes:
        - application/json
      operationId: areas_and_buildings_level_occupation_schedule_add
      parameters:
        - in: body
          name: Occupation
          description: The Occupation to add.
          schema:
            $ref: "#/definitions/Occupation_schedule"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"
  /occupation/areas_and_buildings_level_occupation_schedule_get:
      x-swagger-router-controller: area_and_buildings
      get:
        summary: Get the schedule (Type 0 = Building; Type 1 = Area).
        tags:
                - occupation
        consumes:
          - application/json
        operationId: areas_and_buildings_level_occupation_schedule_get
        parameters:
          - in: query
            name: name
            description: The building or area to receive the value.
            required: true
            type: string
          - in: query
            name: type
            description: (Type 0 = Building; Type 1 = Area).
            required: true
            type: string
        responses:
          200:
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/HelloWorldResponse"
          503:
            description: Error
            schema:
              $ref: "#/definitions/HelloWorldResponse"
...
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  Top_visits:
    properties:  
        username:
          type : string
        token:
          type : string
        area_name:
          type : string
        top:
          type : string 
        building_name:
          type : string
        type:
          type : string  
  Occupation:
    properties:  
        username:
          type : string
        token:
          type : string
        area_name:
          type : string
        occupation:
          type : string
        waiting_time:
          type : string
        building_name:
          type : string
        type:
          type : string
  Occupation_schedule:
    properties:  
        username:
          type : string
        token:
          type : string
        area_name:
          type : string
        occupation:
          type : string
        schedule:
          type : string  
        building_name:
          type : string
        type:
          type : string
...